[RWKV7] Drop some kernels to enhance speed (#346)
* [RWKV] remove unnecessary kernels

---------

+++ tests/ops/test_rwkv7.py
@@ -15,7 +15,7 @@ from fla.utils import device, is_intel_alchemist
 @pytest.mark.parametrize("T", [1024])
 @pytest.mark.parametrize("n_embd", [512, 1024])
 @pytest.mark.parametrize("dim_ffn", [2048, 4096])
-@pytest.mark.parametrize("dtype", [torch.float16])
+@pytest.mark.parametrize("dtype", [torch.bfloat16])
 @pytest.mark.parametrize("inplace", [True, False])
 @pytest.mark.parametrize("xprevdim", [2, 3])
 @pytest.mark.skipif(
